+++ R/retrieve_valuelabel.R
@@ -5,6 +5,7 @@
 #' @param variablename name of the variable 
 #' @param valuelabel "*" (default) retrieves all value labels. 
 #' @param genesis to authenticate a user and set the database (see below).
+#' @param language retrieve information in German "de" (default) or in English "en" if available. 
 #' @param ... other arguments send to the httr::GET request. 
 #'   
 #' @details  
@@ -31,7 +32,7 @@
 retrieve_valuelabel <- function(
 	variablename, 
 	valuelabel="*", 
-	genesis=NULL, ... ) {
+	genesis=NULL, language='de', ... ) {
 
 	genesis <- make_genesis(genesis)
 
@@ -47,7 +48,7 @@ retrieve_valuelabel <- function(
 		kriterium = '',
 		bereich = 'Alle',
 		listenLaenge = '25000',
-		sprache = 'de')
+		sprache = language)
 
 	datenaufbau <- GET(baseurl, query  = param, ... ) 
 	datenaufbau <- content(datenaufbau, type='text/xml', encoding="UTF-8")